[INLONG-5520][Dashboard] Move cluster config to metas
