+++ inlong-dashboard/src/metas/clusters/common/types.ts
